Create archiveretention.sh
Bash script to archive files and then copy new ones.

This script will be used for monthly updates of some documents uploaded to a folder, the old documents need to be archived into one compressed file and the new documents, which may have different names, copied over the top of the old. The script needs to be called on a document by document case from another script. The basic flow for this script should be -

    The script file should create a new gzip archive with a specified name (created from a prefix constant in the script and the current month and year e.g. prefix.september.2009.tar.gz) only if it does not already exist, otherwise add to the existing one.
    Copy the old file into the archive.
    Replace the old file with the new one